v0.0.1 - Uploaded files, logo modal
Uploaded some files to start the work (sourced from the previous projects - Cookbook and Color Palette). Removed most of the unnecessary stuff.

Added a "Logo Modal" - it will be the place where the user can drop the logo. After that, the modal will move out of the window view and the page content will move in.