v0.0.2 - Code cleaning, Navigation, Drop Logo
Improved code (dark mode script + other minor changes).

Implemented working navigation (scroll to the page elements) - burger menu.
The list of nav links is automatically sourced from the page sections.
There is an indicator for the closest section on the page (so the user knows where they are).

Added a drop-file system to take the file and insert it into the page elements.