v0.1.0 - Logo Modal, General CSS, Navigation
Added content to the Logo Modal section.

Added possibility of changing the uploaded logo.
When the user uploads the logo, they can click on the icon (navigation) and the logo modal will smoothly appear, so the user can upload another logo.

When the user changes their mind, they can still come back to the main page without uploading new logo.

Improved and fixed the navigation. When scrolled down the page, the navigation is a little bit more visible.
Now, it won't have a fixed position on mobile (too much clutter).
On smaller devices, it now looks more consistent with the rest of the page.

Made some other CSS changes.