v0.2.0 - Fade-in animation, color value, bug fix
Added a fade-in animation for the page sections. Now, the sections will fade in and out of the window view as the user is scrolling.

Added a color value calculator. It takes the image (logo), makes a lower-quality copy (much better performance) and calculates the average color, lightness and transparency. It is used to determine if the logo background should be darker or lighter,
For example, the user uses light mode and has a mostly white logo. The background behind the uploaded logo will automatically get darker, so the user can see their logo. Similar behavior is applied to a dark logo when using the dark mode.

Fixed a bug where sometimes on page load the logo modal would not be visible, instead the page content (sections, nav, etc.) would be visible.
Now, there's a class added to the body element after the code execution. The page content position (out of the window view) is set only with that class.

Changed some of the CSS code to better fit different screen sizes (logo modal).

Started making a "balance" section - a grid that helps visualize the symmetry of the logo.

Navigation looks broken now, but it's a simple fix. Have to make more page sections and see if the section title looks better in gradient colors or simple gray.