v0.2.1 - Bug fixes, readme, page toggle, logo bg
Fixed a bug where the average color and lightness would not be calculated correctly.
When the image was semi-transparent, the alpha color was read as black. This is why the average color and lightness were darker if a part of the logo was transparent.
Now, only the non-fully transparent pixels are taken into account when calculating these values.

Additionally, the logo background color was adjusted for better visibility (based on the logo lightness).

Added some content to the readme file. More content will be added as the work progresses.

Added a page toggle feature. After the user uploads the logo, they can navigate to the logo modal (left arrow), to the page content (right arrow) or toggle between them (enter).

Made other CSS changes (including the balance section).