v0.3.0 - Accessibility, logo BG color (slider)
Added accessibility features.
The user can use tab to navigate through the program. When the nav link is on focus, the navbar (burger) will open and close if the nav links are out of focus.

Removed page switch on enter click (incompatible with the tab nav), but the left and right arrow switch stays.

Changed the transition duration on some of the elements (it was too long in comparison to the other elements).

Added a possibility of the logo background change. The user can use a range slider to change the bg color (grayscale). The current bg color value is displayed in percent (people wouldn't understand 0-255, so this is the best solution for the current color indication).

The user can click a button to reset the bg color to the standard one.
The automatic bg color adjustment is still present, it just can be overwritten by the user input if necessary.

Fixed a small bug where (if loaded directly into the page content) the navigation wouldn't show the closest section (it would show the last one as the closest).
It wouldn't affect the UX since the logo modal is displayed first, but it's better to fix that.