v0.4.0 - Scalability section
Added a scalability section.
It consists of 2 parts.

The first one is a simple logo display in different sizes (small, medium, large).

In the second part, the user can manually adjust the logo size (width) using a range slider.
The minimum value is set to 24px and the max value is automatically adjusted based on the logo aspect ratio, container width and max container height (so the user doesn't have to scroll 4 screen heights when uploading a vertical logo - edge case, but worth covering).
When the screen is resized, the logo width is adjusted automatically.