v0.6.0 - Pixelated Section, Background Fix
Added a pixelated logo section.
There are 3 versions of the logo, each with a different level of pixelation.

Fixed a bug where the background (image and gradient) would not appear for a short time, because they were in a display:none; parent container. Now, the way they are displayed is changed.
Additionally, the bug fix allowed for setting a longer time to add the class responsible for fixing another bug where the page content (sections) was sometimes visible on page load.