v1.0.0 - Comparison Section, Optimization, Nav
Added the Logo Comparison section.
It contains 11 popular logos and the uploaded one among them (12 total).
The logo placement is generated randomly.

Improved navigation - Now it can't be longer than the page height.
Changed the link to the closest section to a more distinctive color.
Removed the closest section highlight on mobile.
Now, the navigation will hide when the page is scrolled (mobile only).
Navigation has improved visuals (fade in and out animations) - The opacity is 100%, but was gradually transitioning to transparency when hiding.
Added a scrollbar to the nav element.

Improved the Lighthouse score.
Changed some scripts, added alt tags, made some other minor changes.

Added a functionality to the fade-in section animation (script). It calls the function one more time after resizing the window, so it fixed a bug where the section wouldn't fade in when resizing the window really fast.

Added an element check to the scalability script (so there's no console error).

Finished the readme file.

Made some other minor changes.